Add code to force empty modes to be accessible
Previously an empty mode would result in the mode not existing in the
generated code as without callbacks for a mode the mode doesn't exist
when Gremlin is activated. This fix adds an empty non triggerable action
to each mode which ensures that the mode will appear when Gremlin is
activated.