Always run auto release actions, even if mode hasn't changed
Not running the release when the mode hasn't changed makes the
assumption that auto release is only useful to mode change related auto
release. As it is also useful for shift conditions etc. the optimization
of only running the release action once a mode changed is not suitable
for generic use and thus has been removed.