Fixes macros recording a pause of 0 seconds before first axis input
When a macro was recording input for joystick axes, the first input event for each axis had a pause of 0 seconds before it and ignored any wait between a previous action and moving the joystick.

This issue was caused by the behavior when checking if a new JoystickAction should be created. Before recording the action, there are checks to see if the event is outside the polling rate and minimum change amount allowed, but before these checks there is a null check that confirms the event (or more accurately an earlier event with the same hash value) is already in the _recording_times dictionary. If the event is not in the dictionary then an entry is added for this new event and the time for the entry is set to now; this behavior is incorrect because it sabotages the creation of the PauseAction in the next if statement block. The PauseAction created will record the time between now and the last action recorded in the dictionary, which was just added now so there is a zero second pause recorded between the "last" event and this new one because they are the same event.

Adding the new event to the _recording_times dictionary is not even necessary at this step because if the "if event not in self._recording_times" case is true then the "if add_new_entry" will always execute and that code block does the same operation to add the event to _recording_times. Because adding the event to _recording_times will be done later, I have rearranged the if statements such that if the event IS in _recording_times then the checks for polling rate and minimum change will be performed, but if the event is not in _recording_times then we skip the check because this is the very first event from that device and input identifier so there is no need and no way to check the event against the polling rate and minimum change amount.